JMS Formatting & Bank/Broker fix
- Formatted the JMS logic to a MessageQueue class to enforce the DRY principles.
- Bank & Broker not working fixed.
