+++ src/applications/bank/JMSBankFrame.java
@@ -1,5 +1,6 @@
 package applications.bank;
 import mix.messaging.GLOBALS;
+import mix.messaging.MessageQueue;
 import mix.messaging.requestreply.RequestReply;
 import mix.model.bank.BankInterestReply;
 import mix.model.bank.BankInterestRequest;
@@ -35,21 +36,6 @@ public class JMSBankFrame extends JFrame {
 	private JPanel contentPane;
 	private JTextField tfReply;
 	private DefaultListModel<RequestReply<BankInterestRequest, BankInterestReply>> listModel = new DefaultListModel<RequestReply<BankInterestRequest, BankInterestReply>>();
-
-	/**
-	 * JMS Variables
-	 */
-	private Connection connection = null;
-	private Session session = null;
-
-	private Destination sendDestination = null;
-
-	private MessageProducer producer = null;
-	private MessageConsumer consumer = null;
-
-	// Queues
-	private Queue bankInterestReplyQueue = null;
-	private Queue bankInterestRequestQueue = null;
 	
 	/**
 	 * Launch the application.
@@ -129,7 +115,7 @@ public class JMSBankFrame extends JFrame {
 					reply.setLoanRequest(request);
 					System.out.println("BROKER: Get Reply: " + reply);
 
-					produce(reply);
+					new MessageQueue().produce(reply);
 					System.out.println("BROKER: Reply has been Send!");
 				}
 			}
@@ -156,7 +142,7 @@ public class JMSBankFrame extends JFrame {
 		 * 11. The Client consumes the LoanReply.
 		 * ======================================================================================
 		 */
-		consume(GLOBALS.bankInterestRequestQueue, new MessageListener() {
+		new MessageQueue().consume(GLOBALS.bankInterestRequestQueue, new MessageListener() {
 			@Override
 			public void onMessage(Message msg) {
 				if(msg instanceof ObjectMessage) {
@@ -179,88 +165,4 @@ public class JMSBankFrame extends JFrame {
 			}
 		});
 	}
-
-	/**
-	 * This method produces a JMS message and sends it to the Queue.
-	 * @param obj
-	 */
-	private void produce(Serializable obj) {
-		try {
-			if(connection == null) {
-				openJMSConnection();
-			}
-
-			if (obj instanceof  BankInterestReply) {
-				producer = session.createProducer(bankInterestReplyQueue);
-			}
-
-			System.out.println("BANK: ObjectMessage Created");
-			ObjectMessage msg = session.createObjectMessage(obj);
-			producer.send(msg);
-			System.out.println("BANK: ObjectMessage Send");
-		} catch (JMSException e) {
-			e.printStackTrace();
-		} finally {
-			if (connection != null) {
-				try {
-					connection.close();
-				} catch (JMSException e) {
-					e.printStackTrace();
-				}
-			}
-		}
-	}
-
-	/**
-	 * This method consumes a JMS message and handles it's request.
-	 *
-	 * @param queue
-	 * @param listener
-	 */
-	private void consume(String queue, MessageListener listener) {
-		try {
-			if(connection == null) {
-				openJMSConnection();
-			}
-
-			if (queue.equals(GLOBALS.bankInterestRequestQueue)) {
-				System.out.println("BANK: INSIDE BankInterestRequestQueue");
-				consumer = session.createConsumer(bankInterestRequestQueue);
-				System.out.println("BANK: Consumer has been Set!");
-			}
-
-			System.out.println("BANK: After BankInterestRequestQueue");
-			consumer.setMessageListener(listener);
-			System.out.println("BANK: Listener has been set!");
-			connection.start();
-			System.out.println("BANK: Connection has been started!");
-		} catch (JMSException e) {
-			e.printStackTrace();
-		}
-	}
-
-	/**
-	 * This method initiates the JMS connection for the Queue's.
-	 * It establishes an ActiveMQ connection and the corresponding Queue's.
-	 */
-	private void openJMSConnection() {
-		try {
-			ActiveMQConnectionFactory factory = new ActiveMQConnectionFactory();
-
-			// Trust all serializable classes
-			factory.setTrustAllPackages(true);
-
-			connection = factory.createConnection();
-			session = connection.createSession(false, Session.AUTO_ACKNOWLEDGE);
-
-			bankInterestRequestQueue = session.createQueue(GLOBALS.bankInterestRequestQueue);
-			bankInterestReplyQueue = session.createQueue(GLOBALS.bankInterestReplyQueue);
-
-			System.out.println("BANK: Ready to connect!");
-
-		} catch (JMSException e) {
-			e.printStackTrace();
-		}
-	}
-
 }
